fix(icon-button): remove tooltip and add alt text
* applying the `alt` prop to the button's aria-label
* updating examples of suggested alt text
* removing tooltip from button
* consolidating interactive-icon into icon-button
